requirements.txt
# local package
-e .

# external requirements
pyparsing
jira
gitpython
github3.py
javalang
pandas
junitparser
click

# testing requirements
pytest
pytest_steps